+++ src/PaymentMethods/Giftcards.php
@@ -1,0 +1,108 @@
+<?php declare(strict_types=1);
+
+namespace Buckaroo\Shopware6\PaymentMethods;
+
+use Buckaroo\Shopware6\Handlers\GiftcardsPaymentHandler;
+
+class Giftcards implements PaymentMethodInterface
+{
+    /*
+    * @return string
+    */
+    public function getBuckarooKey(): string
+    {
+        return 'giftcards';
+    }
+
+    /**
+     * @return string
+     */
+    public function getVersion(): string
+    {
+        return '2';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getName(): string
+    {
+        return 'Buckaroo Giftcards';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getDescription(): string
+    {
+        return 'Pay with Giftcards';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getPaymentHandler(): string
+    {
+        return GiftcardsPaymentHandler::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string|null
+     */
+    public function getTemplate(): ?string
+    {
+        return null;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getMedia(): string
+    {
+        return __DIR__  . '/../Resources/views/storefront/buckaroo/logo/giftcards.png';
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return array
+     */
+    public function getTranslations(): array
+    {
+        return [
+            'de-DE' => [
+                'name'        => $this->getName(),
+                'description' => 'Bezahlen mit Giftcards',
+            ],
+            'en-GB' => [
+                'name'        => $this->getName(),
+                'description' => $this->getDescription(),
+            ],
+        ];
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return string
+     */
+    public function getType(): string
+    {
+        return 'direct';
+    }
+
+    public function canRefund(): bool
+    {
+        return true;
+    }
+}